backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.22.0
pymongo[srv]>=4.3.3
motor>=3.1.2
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
aiohttp>=3.8.5
msal>=1.23.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
pytest>=7.3.1
python-multipart
beanie>=1.10.0
asyncio>=3.4.3
openai==0.28.1
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
